Split src files, expose functions
